Adds support for the HyStart algorithm to TCP with CUBIC congestion control.

HyStart addresses a common problem during slow-start on paths with high bandwidth delay product (BDP), where using only dropped packets as an indication of congestion detects the congestion too late, and the congestion window (cwnd) has vastly overshot the usable bandwidth of the path, causing many many lost packets.

This implementation of HyStart is based on the Linux Kernel, with some additional commentary on [constant values taken from the HyStart++ RFC](https://www.rfc-editor.org/rfc/rfc9406.html#section-4.3) (note we do not implement the RFC algorithm, preferring to follow Linux).

I have tested this implementation via an embedded Tailscale data plane on a path from Abu Dhabi to Helsinki (140 ms RTT).

Throughput vs time without HyStart shows a big dip around 2 seconds, due to packet loss from the mechanism described above.

![throughput-main-10s](https://github.com/google/gvisor/assets/5375600/5273cc42-fb96-4bfc-95bd-afe62a129821)


With HyStart, slow-start exits just before we start exceeding the path bandwidth, and we stabilize throughput without packet loss.

![throughput-hystart-10s](https://github.com/google/gvisor/assets/5375600/1af8eaaf-6556-4dd3-882c-ff2b7ead35ad)


Over a long-ish upload, this doesn't translate to a huge difference, because even without hystart, we stabilize to the right Tx rate within a few seconds.  But, it makes a big difference for medium size transfers.  If I limit the transfer to 3s (~30MB for my connection):

without HyStart

```
INTERVAL       THROUGHPUT
0.00-1.14 sec  44.1423 Mbits/sec
1.14-2.51 sec  48.9794 Mbits/sec
2.51-3.27 sec  21.9600 Mbits/sec
----------------------------------
0.00-3.27 sec  40.9907 Mbits/sec
```

with HyStart

```
INTERVAL       THROUGHPUT
0.00-1.13 sec  44.7275 Mbits/sec
1.13-2.31 sec  84.6852 Mbits/sec
2.31-3.10 sec  85.4545 Mbits/sec
----------------------------------
0.00-3.10 sec  70.3722 Mbits/sec
```