@hbhasker 2024-04-16
This is great btw. I didn't realize anyone used the cubic in gvisor. Gvisor itself still uses reno by default because most folks don't care about CUBIC in data center environments. Thanks for implementing this! 

Not on the team anymore but it's great to see gvisor/net stack usage and improvements.