@kevinGC 2024-04-16
@hbhasker is there a reason _not_ to make CUBIC default? We'll need to benchmark to confirm, but it seems like kind of a layup given it's implemented already.