@hbhasker 2024-04-16
The benchmarks never showed cubic doing any better at <5ms latency which is majority of gvisor usage.