@hbhasker 2024-04-17
There was another reason. I implemented cubic with floating point operations. It might make sense to change it to fixed decimal similar to how linux does to save CPU costs and avoid floating point errors.