@spikecurtis 2024-04-17
I ran the Iperf benchmarks `"-test.benchtime=30s -test.count=3"` --- just eyeballing the numbers I don't think there are any statistically significant differences.  I picked the numbers somewhat arbitrarily since I don't really know what the different test cases are doing...

[hystart-cubic-bench.txt](https://github.com/google/gvisor/files/15008497/hystart-cubic-bench.txt)
[main-cubic-bench.txt](https://github.com/google/gvisor/files/15008499/main-cubic-bench.txt)