@spikecurtis 2024-04-19
here's what I got for 10 samples:

```
% go run golang.org/x/perf/cmd/benchstat@latest main-cubic-bench10.txt hystart-cubic-bench10.txt
go: downloading golang.org/x/perf v0.0.0-20240404204407-f3e401e020e4
go: downloading github.com/aclements/go-moremath v0.0.0-20210112150236-f10218a38794
goos: linux
goarch: amd64
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
                                                                 │ main-cubic-bench10.txt │      hystart-cubic-bench10.txt       │
                                                                 │         sec/op         │    sec/op     vs base                │
Iperf/operation.Upload-8                                                     717.6n ±  2%   690.1n ±  2%   -3.83% (p=0.001 n=10)
Iperf/operation.Download-8                                                   503.1n ±  2%   501.1n ±  4%        ~ (p=0.280 n=10)
IperfParameterized/operation.Upload/length.4K/parallel.1-8                   2.447µ ±  2%   2.317µ ±  5%   -5.29% (p=0.000 n=10)
IperfParameterized/operation.Upload/length.64K/parallel.1-8                  716.2n ±  3%   739.6n ±  1%   +3.26% (p=0.001 n=10)
IperfParameterized/operation.Upload/length.1024K/parallel.1-8                542.8n ±  2%   547.5n ±  3%   +0.88% (p=0.015 n=10)
IperfParameterized/operation.Upload/length.4K/parallel.16-8                  19.02µ ±  3%   19.26µ ±  3%        ~ (p=0.247 n=10)
IperfParameterized/operation.Upload/length.64K/parallel.16-8                 6.590µ ± 12%   6.662µ ± 12%        ~ (p=0.971 n=10)
IperfParameterized/operation.Upload/length.1024K/parallel.16-8               5.000µ ±  6%   5.661µ ± 11%  +13.22% (p=0.000 n=10)
IperfParameterized/operation.Download/length.4K/parallel.1-8                 1.786µ ±  2%   1.749µ ±  2%        ~ (p=0.052 n=10)
IperfParameterized/operation.Download/length.64K/parallel.1-8                510.0n ±  2%   518.0n ±  1%        ~ (p=0.305 n=10)
IperfParameterized/operation.Download/length.1024K/parallel.1-8              497.3n ±  6%   484.7n ±  4%        ~ (p=0.123 n=10)
IperfParameterized/operation.Download/length.4K/parallel.16-8                8.508µ ±  1%   8.499µ ±  2%        ~ (p=0.699 n=10)
IperfParameterized/operation.Download/length.64K/parallel.16-8               7.551µ ±  1%   7.425µ ±  1%   -1.67% (p=0.000 n=10)
IperfParameterized/operation.Download/length.1024K/parallel.16-8             7.763µ ±  1%   7.598µ ±  2%   -2.13% (p=0.008 n=10)
geomean                                                                      2.162µ         2.165µ         +0.14%

                                                                 │   main-cubic-bench10.txt   │             hystart-cubic-bench10.txt              │
                                                                 │ bandwidth.bytes_per_second │ bandwidth.bytes_per_second  vs base                │
Iperf/operation.Upload-8                                                         1.471G ±  2%                 1.530G ±  2%   +4.01% (p=0.000 n=10)
Iperf/operation.Download-8                                                       2.085G ±  2%                 2.094G ±  4%        ~ (p=0.280 n=10)
IperfParameterized/operation.Upload/length.4K/parallel.1-8                       431.3M ±  2%                 454.6M ±  5%   +5.41% (p=0.000 n=10)
IperfParameterized/operation.Upload/length.64K/parallel.1-8                      1.473G ±  3%                 1.427G ±  1%   -3.09% (p=0.001 n=10)
IperfParameterized/operation.Upload/length.1024K/parallel.1-8                    1.944G ±  2%                 1.930G ±  3%   -0.73% (p=0.035 n=10)
IperfParameterized/operation.Upload/length.4K/parallel.16-8                      56.14M ±  3%                 55.56M ±  3%        ~ (p=0.529 n=10)
IperfParameterized/operation.Upload/length.64K/parallel.16-8                     350.1M ±  8%                 347.7M ± 17%        ~ (p=0.481 n=10)
IperfParameterized/operation.Upload/length.1024K/parallel.16-8                   631.8M ± 33%                 302.8M ± 11%  -52.08% (p=0.000 n=10)
IperfParameterized/operation.Download/length.4K/parallel.1-8                     587.6M ±  3%                 602.0M ±  2%   +2.45% (p=0.029 n=10)
IperfParameterized/operation.Download/length.64K/parallel.1-8                    2.057G ±  2%                 2.026G ±  2%        ~ (p=0.280 n=10)
IperfParameterized/operation.Download/length.1024K/parallel.1-8                  2.113G ±  6%                 2.167G ±  4%        ~ (p=0.143 n=10)
IperfParameterized/operation.Download/length.4K/parallel.16-8                    153.5M ±  7%                 156.4M ± 10%        ~ (p=0.315 n=10)
IperfParameterized/operation.Download/length.64K/parallel.16-8                   175.1M ±  4%                 182.9M ±  3%   +4.50% (p=0.015 n=10)
IperfParameterized/operation.Download/length.1024K/parallel.16-8                 167.7M ±  8%                 177.1M ±  9%   +5.58% (p=0.035 n=10)
geomean                                                                          584.4M                       562.1M         -3.81%

                                                                 │ main-cubic-bench10.txt │       hystart-cubic-bench10.txt       │
                                                                 │          B/s           │      B/s       vs base                │
IperfParameterized/operation.Upload/length.4K/parallel.1-8                  1.559Gi ±  2%   1.647Gi ±  5%   +5.61% (p=0.000 n=10)
IperfParameterized/operation.Upload/length.64K/parallel.1-8                 85.21Gi ±  3%   82.52Gi ±  1%   -3.16% (p=0.001 n=10)
IperfParameterized/operation.Upload/length.1024K/parallel.1-8               1.757Ti ±  2%   1.742Ti ±  3%   -0.87% (p=0.015 n=10)
IperfParameterized/operation.Upload/length.4K/parallel.16-8                 205.3Mi ±  3%   202.8Mi ±  3%        ~ (p=0.247 n=10)
IperfParameterized/operation.Upload/length.64K/parallel.16-8                9.262Gi ± 11%   9.168Gi ± 11%        ~ (p=0.971 n=10)
IperfParameterized/operation.Upload/length.1024K/parallel.16-8              195.3Gi ±  6%   172.5Gi ± 10%  -11.66% (p=0.000 n=10)
IperfParameterized/operation.Download/length.4K/parallel.1-8                2.136Gi ±  3%   2.182Gi ±  2%        ~ (p=0.052 n=10)
IperfParameterized/operation.Download/length.64K/parallel.1-8               119.7Gi ±  2%   117.8Gi ±  1%        ~ (p=0.315 n=10)
IperfParameterized/operation.Download/length.1024K/parallel.1-8             1.918Ti ±  6%   1.967Ti ±  4%        ~ (p=0.123 n=10)
IperfParameterized/operation.Download/length.4K/parallel.16-8               459.1Mi ±  1%   459.6Mi ±  2%        ~ (p=0.684 n=10)
IperfParameterized/operation.Download/length.64K/parallel.16-8              8.084Gi ±  1%   8.220Gi ±  1%   +1.69% (p=0.000 n=10)
IperfParameterized/operation.Download/length.1024K/parallel.16-8            125.8Gi ±  1%   128.5Gi ±  2%   +2.17% (p=0.009 n=10)
geomean                                                                     22.81Gi         22.69Gi         -0.52%
```